backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "httpx>=0.28.1",
    "kafka-python-ng>=2.2.3",
    "psycopg[binary]>=3.2.4",
]
